adding basics for setting the curent user id to cache and playing with goals

+++ Blastdan.BambooHr.Infrastructure/Models/GoalsCache.cs
@@ -18,6 +18,6 @@ namespace Blastdan.BambooHr.Infrastructure.Models
             this.CurrentUserEmployeeId = employee.EmployeeId;
         }
 
-        public int CurrentUserEmployeeId { get; set; }
+        public long CurrentUserEmployeeId { get; set; }
     }
 }
